Showing an 'Repository is empty' warning on repository/show
The default behavior is to 404 when the repository exists,
but is empty / bare.
However, in that case, the repositories/show template is not rendered,
and thus no clone URLs are displayed.

This commit always renders the repositories/show template,
and additionally displays a warning when the repository is empty.

This may be a solution to #19