Most of the functionality works (with some manual patches in database)
Deployment credentials, Post receive URLs, Repository mirrors and Git config keys work when the configuration is corrected with the functions provided in the Rescue tab of the plugin's settings. The permissions of the repositories in the file system also have to be corrected manually to grant access to both Gitolite and OpenProject.